+++ packages/tiff/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@jimp/tiff",
+  "version": "0.3.5",
+  "description": "Default Jimp tiff encoder/decoder.",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "MIT",
+  "dependencies": {
+    "utif": "^2.0.0"
+  }
+}
